chore: type and enum housekeeping
