+++ libs/application/templates/accident-notification/src/fields/ApplicationStatus/index.tsx
@@ -14,7 +14,7 @@ import { useCallback, useEffect } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { getAccidentStatusQuery } from '../../hooks/useLazyStatusOfNotification'
 import { inReview } from '../../lib/messages'
-import { ReviewApprovalEnum, SubmittedApplicationData } from '../../types'
+import { ReviewApprovalEnum, SubmittedApplicationData } from '../../utils/types'
 import { isUniqueAssignee } from '../../utils/miscUtils'
 import { StatusStep } from './StatusStep'
 import { ApplicationStatusProps } from './StatusStep/types'
