+++ libs/application/templates/accident-notification/src/fields/FormOverviewInReview/index.tsx
@@ -5,9 +5,9 @@ import { Box, Button, Divider, Text } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
 import { InputController } from '@island.is/shared/form-fields'
 import { inReview, overview, thirdPartyComment } from '../../lib/messages'
-import { ReviewApprovalEnum } from '../../types'
 import { FormOverview } from '../FormOverview'
 import { ConfirmationModal } from './ConfirmationModal'
+import { ReviewApprovalEnum } from '../../utils/enums'
 
 type Props = {
   field: {
