+++ libs/application/templates/accident-notification/src/forms/AccidentNotificationForm/aboutTheAccidentSection/locationSubSection.ts
@@ -21,7 +21,7 @@ import {
   ProfessionalAthleteAccidentLocationEnum,
   RescueWorkAccidentLocationEnum,
   StudiesAccidentLocationEnum,
-} from '../../../types'
+} from '../../../utils/types'
 import {
   isAgricultureAccident,
   isFishermanAccident,
