+++ libs/application/templates/accident-notification/src/index.ts
@@ -1,12 +1,13 @@
 import AccidentNotificationTemplate from './lib/AccidentNotificationTemplate'
 import { AccidentNotification } from './lib/dataSchema'
 import * as appMessages from './lib/messages'
-import { OnBehalf } from './types'
+import { OnBehalf } from './utils/enums'
+
 import * as appUtils from './utils/miscUtils'
 
 export const getFields = () => import('./fields')
 
-export * from './types'
+export * from './utils/types'
 export * from './shared'
 export { OnBehalf }
