+++ libs/application/templates/accident-notification/src/shared/constants.ts
@@ -5,7 +5,7 @@ import {
   ProfessionalAthleteAccidentLocationEnum,
   RescueWorkAccidentLocationEnum,
   StudiesAccidentLocationEnum,
-} from '../types'
+} from '../utils/enums'
 
 export enum ApiActions {
   submitApplication = 'submitApplication',
