+++ libs/application/templates/accident-notification/src/utils/constants.ts
@@ -1,0 +1,3 @@
+export const UPLOAD_ACCEPT = '.pdf, .doc, .docx, .rtf, .jpg, .jpeg, .png, .heic'
+
+export const FILE_SIZE_LIMIT = 10000000 // 10MB
