+++ libs/application/templates/accident-notification/src/utils/documentUtils.spec.ts
@@ -2,11 +2,6 @@ import { FormatMessage } from '@island.is/localization'
 
 import { AccidentNotification } from '../lib/dataSchema'
 import {
-  AttachmentsEnum,
-  PowerOfAttorneyUploadEnum,
-  WhoIsTheNotificationForEnum,
-} from '../types'
-import {
   getAttachmentTitles,
   isUploadNow,
   returnMissingDocumentsList,
@@ -18,6 +13,11 @@ import {
 } from './documentUtils'
 import { FormValue } from '@island.is/application/types'
 import { YesOrNoEnum } from '@island.is/application/core'
+import {
+  AttachmentsEnum,
+  PowerOfAttorneyUploadEnum,
+  WhoIsTheNotificationForEnum,
+} from './enums'
 
 describe('hasMissingDocuments', () => {
   it('should return true when missing documents', () => {
@@ -58,13 +58,19 @@ describe('hasReceivedAllDocuments', () => {
   const testCases = [
     { who: WhoIsTheNotificationForEnum.ME, fatal: YesOrNoEnum.NO },
     { who: WhoIsTheNotificationForEnum.JURIDICALPERSON, fatal: YesOrNoEnum.NO },
-    { who: WhoIsTheNotificationForEnum.POWEROFATTORNEY, fatal: YesOrNoEnum.YES },
+    {
+      who: WhoIsTheNotificationForEnum.POWEROFATTORNEY,
+      fatal: YesOrNoEnum.YES,
+    },
     { who: WhoIsTheNotificationForEnum.POWEROFATTORNEY, fatal: YesOrNoEnum.NO },
   ]
   it.each(testCases)(
     'should return true when all documents are received',
     (data) => {
       const answers = getNoMissingDocuments() as AccidentNotification
+      if (!answers.whoIsTheNotificationFor) {
+        return
+      }
       answers.whoIsTheNotificationFor.answer = data.who
       answers.wasTheAccidentFatal = data.fatal
       expect(hasReceivedAllDocuments(answers)).toEqual(true)
@@ -73,6 +79,9 @@ describe('hasReceivedAllDocuments', () => {
 
   it.each(testCases)('should return false when missing documents', (data) => {
     const answers = getMissingDocuments() as AccidentNotification
+    if (!answers.whoIsTheNotificationFor) {
+      return
+    }
     answers.whoIsTheNotificationFor.answer = data.who
     answers.wasTheAccidentFatal = data.fatal
     expect(hasReceivedAllDocuments(answers)).toEqual(false)
