+++ libs/application/templates/accident-notification/src/utils/getOptions.spec.ts
@@ -1,6 +1,6 @@
 import { FormValue } from '@island.is/application/types'
-import { AccidentTypeEnum, WhoIsTheNotificationForEnum } from '../types'
 import { getAccidentTypeOptions } from './getOptions'
+import { AccidentTypeEnum, WhoIsTheNotificationForEnum } from './enums'
 
 describe('getAccidentTypeOptions', () => {
   const onBehalfOfEmployee: FormValue = {
